pyproject.toml
[tool.poetry]
name = "brotherhoods-backend"
version = "0.1.0"
description = ""
authors = ["macwdo <danilo.macedofernandes@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.3"
django-rest-framework = "^0.1.0"
ipdb = "^0.13.13"
ruff = "^0.3.3"
drf-nested-routers = "^0.93.5"
djangorestframework-simplejwt = "^5.3.1"
pydantic = "^2.6.4"
pytest = "^8.1.1"
pytest-django = "^4.8.0"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.7"}
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.14.5"}
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
coverage = "^7.4.4"
pre-commit = "^3.7.0"
django-cors-headers = "^4.3.1"
freezegun = "^1.4.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "project.settings"
# -- recommended but optional:
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


